Adapt examples to optimizations Reestablish editable steps

+++ dff/simulation/steps/custom_input.py
@@ -5,5 +5,5 @@ def custom_input_prepare_variables(step):
     if step.assignable:
         pattern = tf.Variable(step.pattern, dtype=tf.float32, trainable=step.trainable)
     else:
-        pattern = tf.constant(step.pattern)
+        pattern = tf.constant(step.pattern, dtype=tf.float32)
     return {"pattern": pattern}
